feat: split navigation into separate files
