+++ src/components/index.ts
@@ -2,6 +2,6 @@ export { CodeBlock } from './CodeBlock'
 export { Divider } from './Divider'
 export { InlineCode } from './InlineCode'
 export { Main } from './Main'
-export { Navigation } from './Navigation'
+export { Navigation } from './navigation'
 export { PropsTable } from './PropsTable'
 export { Pagination } from './Pagination'
