+++ src/components/navigation/Navigation.tsx
@@ -1,0 +1,104 @@
+import { Heading, styled } from '@atom-learning/components'
+import { capitalCase } from 'capital-case'
+import * as React from 'react'
+
+import { NavigationContent } from './NavigationContent'
+import { NavigationTrigger } from './NavigationTrigger'
+
+export type PageData = {
+  id: string
+  title: string
+  source: 'overview' | 'theme' | 'components'
+}
+
+export type NavigationItem =
+  | PageData[]
+  | {
+      [key: string]: PageData[]
+    }
+
+type NavigationProps = {
+  items: {
+    [key: string]: NavigationItem
+  }
+}
+
+const useOnClickOutside = (ref, handler) => {
+  React.useEffect(() => {
+    const listener = (event) => {
+      if (!ref.current || ref.current.contains(event.target)) {
+        return
+      }
+      handler(event)
+    }
+
+    document.addEventListener('mousedown', listener)
+    document.addEventListener('touchstart', listener)
+
+    return () => {
+      document.removeEventListener('mousedown', listener)
+      document.removeEventListener('touchstart', listener)
+    }
+  }, [ref, handler])
+}
+
+const StyledNavigation = styled('nav', {
+  backgroundColor: 'white',
+  borderRight: '1px solid $tonal300',
+  boxShadow: '$1',
+  height: '100vh',
+  left: 0,
+  overflowX: 'hidden',
+  overflowY: 'auto',
+  p: '$3',
+  position: 'fixed',
+  top: 0,
+  width: 260,
+  transform: 'translateX(-280px)',
+  zIndex: 1,
+  transition: 'transform 175ms ease-out',
+  when: {
+    lg: {
+      boxShadow: 'none',
+      position: 'sticky',
+      transform: 'none',
+      transition: '0ms'
+    }
+  },
+  variants: {
+    open: {
+      true: {
+        transform: 'translateX(0)'
+      }
+    }
+  }
+})
+
+export const Navigation: React.FC<NavigationProps> = ({ items }) => {
+  const ref = React.useRef()
+  const [menuOpen, setMenuOpen] = React.useState(false)
+
+  useOnClickOutside(ref, () => setMenuOpen(false))
+
+  return (
+    <>
+      <NavigationTrigger onClick={() => setMenuOpen(true)} />
+      <StyledNavigation ref={ref} open={menuOpen}>
+        <Heading size="sm" css={{ mb: '$4', width: 200 }}>
+          Atom Learning Design System
+        </Heading>
+        {Object.entries(items).map(([source, content]) => (
+          <React.Fragment key={source}>
+            <Heading as="h2" size="sm" css={{ mb: '$2' }}>
+              {capitalCase(source)}
+            </Heading>
+            <NavigationContent
+              content={content}
+              onNavigate={() => setMenuOpen(false)}
+            />
+          </React.Fragment>
+        ))}
+      </StyledNavigation>
+    </>
+  )
+}
